Move Planner code to helper class
Move Planner code to helper class.  Also remove no longer needed usings from Program.cs.
